final WSDDN code

+++ object_detection_networks/WSDDN/wsddn.py
@@ -307,7 +307,7 @@ class WeaklySupervisedDetection(tf.keras.Model):
 
             #creating filter for bounding box regions with low IoU score
             region_ious = bounding_box_iou(other_score_region, tf.tile(tf.expand_dims(highest_score_region, axis=0),[num_regions-1,1]) )
-            region_ious = tf.cast(region_ious>hp.spatial_reg_iou_threshold, dtype=tf.int8)
+            region_ious = tf.cast(region_ious>hp.spatial_reg_iou_threshold, dtype=tf.float32)
             pdb.set_trace()
             #mask out fc7 output for regions with < threshold IoU
             fc7_output_mask = tf.tile(tf.expand_dims(region_ious,axis=1), (1,fc7.shape[1]))
